requirement.txt
-i https://pypi.org/simple
certifi==2019.3.9
chardet==3.0.4
django-common-helpers==0.9.2
django-cron==0.5.1
django==2.1.7
idna==2.8
pyserial==3.4
python-dotenv==0.10.1
python-gsmmodem-new==0.12
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.55.0
